+++ src/index.js
@@ -83,7 +83,8 @@ import {
 	getTimeRemainingUntilLocal,
 	getWeekNumber,
 	isNextDay,
-	isSameDate
+	isSameDate,
+	toDayjs
 } from './services/dateUtils.js';
 
 import {
@@ -278,6 +279,7 @@ import {
 } from './services/user/permissions.js';
 
 import {
+	deleteProfilePicture,
 	otherStats
 } from './services/user/profile.js';
 
@@ -291,6 +293,7 @@ import {
 	createPracticeNotes,
 	deletePracticeSession,
 	deleteUserActivity,
+	fetchRecentActivitiesActiveTabs,
 	findIncompleteLesson,
 	getPracticeNotes,
 	getPracticeSessions,
@@ -332,6 +335,7 @@ export {
 	deletePicture,
 	deletePlaylist,
 	deletePracticeSession,
+	deleteProfilePicture,
 	deleteUserActivity,
 	duplicatePlaylist,
 	editComment,
@@ -399,6 +403,7 @@ export {
 	fetchPlaylist,
 	fetchPlaylistItems,
 	fetchRecent,
+	fetchRecentActivitiesActiveTabs,
 	fetchRecentUserActivities,
 	fetchRelatedLessons,
 	fetchRelatedRecommendedContent,
@@ -528,6 +533,7 @@ export {
 	setupAccount,
 	startLiveEventPolling,
 	status,
+	toDayjs,
 	togglePlaylistPrivate,
 	unEnrollUserInGuidedCourse,
 	unPinGuidedCourse,